+++ docsig/_signature.py
@@ -30,7 +30,7 @@ class Signature:
         self._arguments = arguments
         self._args = _Params()
         self._args.extend(
-            _Param(self._param, a.name, None)
+            _Param(self._param, a.name, None, 0)
             for a in self._arguments.args
             if not _isprotected(a.name)
         )
@@ -45,17 +45,17 @@ class Signature:
     def _get_args_kwargs(self) -> None:
         vararg = self._arguments.vararg
         if vararg is not None and not _isprotected(vararg):
-            self._args.append(_Param(self._param, f"*{vararg}", None))
+            self._args.append(_Param(self._param, f"*{vararg}", None, 0))
 
         if self._arguments.kwonlyargs:
             self._args.extend(
-                _Param(self._param, k.name, None)
+                _Param(self._param, k.name, None, 0)
                 for k in self._arguments.kwonlyargs
             )
 
         kwarg = self._arguments.kwarg
         if kwarg is not None and not _isprotected(kwarg):
-            self._args.append(_Param(self._param, f"**{kwarg}", None))
+            self._args.append(_Param(self._param, f"**{kwarg}", None, 0))
 
     def _get_returns(self, returns: _ast.NodeNG | None) -> str | None:
         if isinstance(returns, _ast.Name):