+++ docsig/messages.py
@@ -16,6 +16,7 @@ docsig.messages
 | E113: function is missing a docstring
 | E114: class is missing a docstring
 | E115: syntax error in description
+| E116: param not indented correctly
 """
 # Exxx: Error
 # E1xx: Docstring
@@ -34,10 +35,10 @@ E112 = "E112: spelling error found in documented parameter"
 E113 = "E113: function is missing a docstring"
 E114 = "E114: class is missing a docstring"
 E115 = "E115: syntax error in description"
+E116 = "E116: param not indented correctly"
 
 # Hxxx: Hint
 # H1xx: Docstring
-H101 = "hint: there appears to be a syntax error that could be causing this"
-H102 = "hint: documentation is sufficient as a getter is the value returned"
-H103 = "hint: it is possible a syntax error could be causing this"
-H104 = "hint: a class does not return a value during instantiation"
+H101 = "hint: documentation is sufficient as a getter is the value returned"
+H102 = "hint: it is possible a syntax error could be causing this"
+H103 = "hint: a class does not return a value during instantiation"